+++ src/Templates/default/header.tpl.php
@@ -3,7 +3,7 @@
 	<head>
 		<title><?php print $title ?></title>
 		<?php foreach($stylesheets as $style): ?>
-			<link rel="stylesheet" type="text/css" href="<?php print $style; ?>" media="all" />
+		<link rel="stylesheet" type="text/css" href="<?php print $style; ?>" media="all" />
 		<?php endforeach; ?>
 	</head>
 	<body class="<?php print $bodyClasses; ?>">