+++ tests/bootstrap.php
@@ -8,12 +8,20 @@
 	 */
 	class MockController extends \Controller {
 		public function index() {
-			$this->children[] = new \Core\ClassName("Header"); 
+			$this->children[] = new \Core\ClassName("MockChild"); 
 		}
 
 		private function privateFunction() {}
 
 		protected function protectedFunction() {}
+	}
+
+	class MockChildController extends \Controller {
+		public static string $testkey = "test";
+
+		public function index() {
+			$this->response->data[self::$testkey] = "hello world";
+		}
 	}
 
 	ini_set("display_errors", 1);